Add basic CMake configuration
Before I can invest in using DDS with its recent changes, enable CMake
so that I can just use something for now.

There is some messy behavior due to the peculiar use of Catch2 because
I'm using the Catch2Config.cmake from Catch2 2.5.0, but using the Catch2
headers for 2.13. This means that it may be necessary to update this
once the 2.13 Config.cmake file can be used.