Add comparison default constructibility
When the comparison helpers have already captured a value, they can be
default constructible if the value they capture is empty and default
constructible. For example, `lt(std::integral_constant<int, 10>())`
already holds all of the information about the value in the type
information; the object is empty. This means that we can discard the
predicate and recreate it from scratch when we need it.

The motivation behind this is to be able to lift the predicate into
constexpr so that it can be passed as a regular parameter, but used in a
constexpr context. For example:

    template <typename P>
    constexpr auto foo(P) -> std::conditional_t<P()(42), ..., ...>;

By discarding the predicate and recreating it in a constexpr context, we
smuggle the information to constexpr:

    foo(lt(std::integral_constant<int, 10>()));  // This works.
    foo(lt(10));  // This doesn't work.

There are other ways in which this can have utility, but this is the
major area.