Add a custom bitset and assert macro
std::bitset is not constexpr friendly, and I want that for some reason.
Probably mostly just for the entertaining exercise of implementing
things like this, but having fun is reason enough for now :-).

This also allows me to fully control the bitset type, giving me
flexibility to play around with the API, should I want to do so when
e.g. optimizing.