Guard against misuse for bitset<0>
Rather than simply doing a static_assert(N != 0), this does a slightly
more complicated thing by disabling functions which are invalid in this
case. This means that `bitset<0>` can still be used in generic code, but
it moves the errors to compile time rather than runtime.

This may be a bit problematic, because it makes `bitset<0>` not have the
same functions as `bitset<N>`, but I think this should generally be
okay.