Add a generic bounded<> type (refinement)
This is a primitive refinement type, allowing the refinement of a type T
to be within the bounds specified by the arguments.

Given `bounded<T, L, H>`, the `T t` that the `bounded` holds must be:
`L <= t && t <= H`. Note the inclusive-inclusive range. The range is NOT
half-open, because a fully-inclusive range makes specifying that the
maximum value of `T` (e.g. INT_MAX) is within the bounds easier;
otherwise, a sentinel would need to be used.